refactor: add route in sidebar components and make a little fix in dashboard
